+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.rydmike.flexcolorscheme
+package com.rydmike.themesplayground
 
 import io.flutter.embedding.android.FlutterActivity